pyproject.toml
[project]
name = "fastmcp-koeln-presse"
version = "1.0.0"
description = "MCP-Server für RSS-Pressemitteilungen der Stadt Köln"
authors = [
    {name = "Senior Python Engineer", email = "engineer@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "lxml>=4.9.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "uvicorn>=0.20.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/koeln_presse"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"